create jupyter notebook
This commit adds a jupyter notebook to work interactively.
Note that this notebook requires YOUR local copy of
`lazyaudio`. For this notebook to work properly,
the `lazyaudio` folder must be on the same directory(top-level).
If working of google colab, you may need to download the
folder separately, and then run the notebook.